Fix ClassNotFoundException for IOStreamHelper, moved the helper class to api under helpers package 'com.newrelic.api.agent.security.instrumentation.helpers'
