+++ instrumentation-security/java-io-stream/src/main/java/java/io/Reader_Instrumentation.java
@@ -12,7 +12,7 @@ import com.newrelic.api.agent.security.instrumentation.helpers.GenericHelper;
 import com.newrelic.api.agent.weaver.MatchType;
 import com.newrelic.api.agent.weaver.Weave;
 import com.newrelic.api.agent.weaver.Weaver;
-import com.newrelic.agent.security.instrumentation.inputstream.IOStreamHelper;
+import com.newrelic.api.agent.security.instrumentation.helpers.IOStreamHelper;
 
 @Weave(type = MatchType.BaseClass, originalName = "java.io.Reader")
 public abstract class Reader_Instrumentation {
