+++ newrelic-security-api/src/main/java/com/newrelic/api/agent/security/instrumentation/helpers/IOStreamHelper.java
@@ -1,4 +1,4 @@
-package com.newrelic.agent.security.instrumentation.inputstream;
+package com.newrelic.api.agent.security.instrumentation.helpers;
 
 import com.newrelic.api.agent.security.NewRelicSecurity;
 import com.newrelic.api.agent.security.instrumentation.helpers.GenericHelper;
